refactor: rename grid file
